Progress check 6
quickSort() is implemented using the same implementation as heapSort(). I deleted the testSortingAlgorithms button since I couldn't figure out why it wasn't working and realistically the button should be used during coding to test but everything is done so it doesn't make sense for it to even be on the screen but I am done so I probably will never touch this project ever again!